[EI-257] [web] refactor: Extract Messages component from Chat component
