+++ apps/web/app/ui/pages/workspace/chat-ai-navigator.tsx
@@ -1,6 +1,6 @@
 'use client';
 import { Sidebar } from 'react-pro-sidebar';
-import Chat from '../../components/domain/linguistic-guidance/chat';
+import - from '../../components/domain/linguistic-guidance/chat';
 
 export default function ChatAiNavigator() {
   return (
